First attempt at changing the login system
Trying to make sure the password is not stored anywhere after logging in. Later I will probably change it from a bool to a authentication ticket so its not simple to mess with. I need to also be able to store the user and level of access in the main function as well as be able to access the MySQL created variables so I don't have to disconnect and reconnect for every time that I need a request