Changed how the login system works
Made it so the connection happens in main. After the login and password are checked and the auth level of user is determined, the connection is closed. The password is no longer stored past the comparison. Statement and result classes are made in the respective functions and are deleted before exiting the function